Moved the Recommendations list to `RecommendationList.tsx`
refs https://github.com/TryGhost/Product/issues/3791
