+++ apps/admin-x-settings/src/components/settings/site/Recommendations.tsx
@@ -1,16 +1,9 @@
-import Avatar from '../../../admin-x-ds/global/Avatar';
 import Button from '../../../admin-x-ds/global/Button';
-import ConfirmationModal from '../../../admin-x-ds/global/modal/ConfirmationModal';
-import NiceModal from '@ebay/nice-modal-react';
 import React, {useState} from 'react';
-import SettingGroup from '../../../admin-x-ds/settings/SettingGroup';
-import Table from '../../../admin-x-ds/global/Table';
-import TableCell from '../../../admin-x-ds/global/TableCell';
-import TableRow from '../../../admin-x-ds/global/TableRow';
-import useRouting from '../../../hooks/useRouting';
 import RecommendationList from './recommendations/RecommendationList';
 import SettingGroup from '../../../admin-x-ds/settings/SettingGroup';
 import TabView from '../../../admin-x-ds/global/TabView';
+import useRouting from '../../../hooks/useRouting';
 import useSettingGroup from '../../../hooks/useSettingGroup';
 import {useBrowseRecommendations} from '../../../api/recommendations';
 
@@ -30,21 +23,6 @@ const Recommendations: React.FC<{ keywords: string[] }> = ({keywords}) => {
     const buttons = (
         <Button color='green' label='Add recommendation' link={true} onClick={() => {
             openAddNewRecommendationModal();
-        }} />
-    );
-  
-    const action = (
-        <Button color='red' label='Remove' link onClick={() => {
-            NiceModal.show(ConfirmationModal, {
-                title: 'Remove recommendation',
-                prompt: <>
-                    <p>Your recommendation <strong>Lenny Nesletter</strong> will no longer be visible to your audience.</p>
-                </>,
-                okLabel: 'Remove',
-                onOk: async (modal) => {
-                    modal?.remove();
-                }
-            });
         }} />
     );
 
@@ -67,9 +45,9 @@ const Recommendations: React.FC<{ keywords: string[] }> = ({keywords}) => {
             description="Share favorite sites with your audience"
             keywords={keywords}
             navid='recommendations'
+            saveState={saveState}
             testId='recommendations'
             title="Recommendations"
-            saveState={saveState}
             onSave={handleSave}
         >
             <TabView selectedTab={selectedTab} tabs={tabs} onTabChange={setSelectedTab} />
